requirements.txt
fastapi
uvicorn
python-dotenv
fastapi-mail
email-validator
python-multipart